implement router.onReady
