--- test/unit/specs/add-routes.spec.js
@@ -1,34 +1,0 @@
-import Router from '../../../src/index'
-
-describe('router.addRoutes', () => {
-  it('should work', () => {
-    const router = new Router({
-      mode: 'abstract',
-      routes: [
-        { path: '/a', component: { name: 'A' }}
-      ]
-    })
-
-    router.push('/a')
-    let components = router.getMatchedComponents()
-    expect(components.length).toBe(1)
-    expect(components[0].name).toBe('A')
-
-    router.push('/b')
-    components = router.getMatchedComponents()
-    expect(components.length).toBe(0)
-
-    router.addRoutes([
-      { path: '/b', component: { name: 'B' }}
-    ])
-    components = router.getMatchedComponents()
-    expect(components.length).toBe(1)
-    expect(components[0].name).toBe('B')
-
-    // make sure it preserves previous routes
-    router.push('/a')
-    components = router.getMatchedComponents()
-    expect(components.length).toBe(1)
-    expect(components[0].name).toBe('A')
-  })
-})
