--- test/unit/specs/navigation-callback.spec.js
@@ -1,70 +1,0 @@
-import Router from '../../../src/index'
-
-describe('Programmatic navigation callbacks', () => {
-  let calls = []
-  let router, spy1, spy2
-
-  const Foo = {
-    beforeRouteEnter (to, from, next) {
-      calls.push(3)
-      setTimeout(() => {
-        calls.push(4)
-        next()
-      }, 1)
-    }
-  }
-
-  beforeEach(() => {
-    calls = []
-    spy1 = jasmine.createSpy('complete')
-    spy2 = jasmine.createSpy('abort')
-
-    router = new Router({
-      routes: [
-        { path: '/foo', component: Foo }
-      ]
-    })
-
-    router.beforeEach((to, from, next) => {
-      calls.push(1)
-      setTimeout(() => {
-        calls.push(2)
-        next()
-      }, 1)
-    })
-  })
-
-  it('push complete', done => {
-    router.push('/foo', () => {
-      expect(calls).toEqual([1, 2, 3, 4])
-      done()
-    })
-  })
-
-  it('push abort', done => {
-    router.push('/foo', spy1, spy2)
-    router.push('/bar', () => {
-      expect(calls).toEqual([1, 1, 2, 2])
-      expect(spy1).not.toHaveBeenCalled()
-      expect(spy2).toHaveBeenCalled()
-      done()
-    })
-  })
-
-  it('replace complete', done => {
-    router.replace('/foo', () => {
-      expect(calls).toEqual([1, 2, 3, 4])
-      done()
-    })
-  })
-
-  it('replace abort', done => {
-    router.replace('/foo', spy1, spy2)
-    router.replace('/bar', () => {
-      expect(calls).toEqual([1, 1, 2, 2])
-      expect(spy1).not.toHaveBeenCalled()
-      expect(spy2).toHaveBeenCalled()
-      done()
-    })
-  })
-})
